[SWP391 - 115] tạo transaction create rental cho đặt trước và trực tiếp, thêm get và update status cho rental

+++ src/main/java/com/swp391/e_Motion_be/entity/Payment.java
@@ -24,7 +24,7 @@ public class Payment {
     @Column(name = "payment_id")
     private Long id;
     @Column(name = "total_amount", nullable = false)
-    private long amount;
+    private double amount;
     @Column(name = "payment_method", nullable = false)
     @Enumerated(EnumType.STRING)
     private PaymentMethod method;
